AdcReader::threadBody( ) is fixed!!!
The rtems thread-launching mechanism is prototyped like this:
	rtems_task_start(rtems_id, void(*taskFunc)(taskarg), taskArg)
We can't use a class' method for taskFunc() 'cause that would have the "this" pointer silently included in the method call.

So to get around this, we can use a static class-method for taskFunc(), as static methods have no "this" ptr in their call. But, a static method cannot access non-static class attributes or methods. So, within the static method, we can call another (private) non-static method as the "actual" taskFunc(), or thread-body, by passing a ptr to the class' "this" within the static method.

Whew.... just look at AdcReader.h and AdcReader.cc (in particular, the threadStart( ) and threadBody( ) definitions).